structural: move the filepath codec out of the library package
SeasonPathMetadataParser started reading folder names through
FilepathCodec, which lived in services.library — and services.library
already depends on services.parsers.show for EpisodePathMetadataParser,
EpisodePathResult, and SeasonPathMetadataParser. The two packages pointed
at each other.

FilepathCodec imports nothing but java.net and java.nio, so it belongs
below both of them. services.filepath is a leaf every caller can depend on
with nothing depending back. The same move unwinds two more cycles of the
same shape: HlsStreamingService and FileProcessingTaskCoordinator each
reached into services.library for this codec and nothing else.

Pure move — package declarations, imports, and the test's package. No
behavior change.
