+++ src/main/java/com/streamarr/server/services/task/FileProcessingTaskCoordinator.java
@@ -3,7 +3,7 @@ package com.streamarr.server.services.task;
 import com.streamarr.server.domain.task.FileProcessingTask;
 import com.streamarr.server.domain.task.FileProcessingTaskStatus;
 import com.streamarr.server.repositories.task.FileProcessingTaskRepository;
-import com.streamarr.server.services.library.FilepathCodec;
+import com.streamarr.server.services.filepath.FilepathCodec;
 import java.lang.management.ManagementFactory;
 import java.net.InetAddress;
 import java.nio.file.Path;
